Client

+++ app/src/main/java/com/example/bkzalo/listeners/CheckAddListener.java
@@ -5,6 +5,6 @@ import com.example.bkzalo.models.UserModel;
 import java.util.List;
 
 public interface CheckAddListener {
-    void onCheckClick(Long id_userModel);
+    void onCheckClick(int id_userModel);
 
 }
